+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,8 +3,10 @@
 ## Unreleased
 
 * Changed
+  * Added test for `terra-slide-panel`.
+  * Updated docs for `terra-slide-group`.
   * Updated doc and added test for terra-collapsible-menu-view
-
+  
 * Added
   * Added package-lock.json